+++ api/helpers/errorHandler.js
@@ -1,7 +1,20 @@
-const createError = require('http-errors');
+const errorHandler = (error) => {
+	if (error.syscall !== 'listen') {
+		throw error;
+	}
 
-const errorHandler = (next, errorCode) => {
-    next(createError(errorCode));
-}
+	const bind = typeof port === 'string' ? 'Pipe ' + port : 'Port ' + port;
+
+	switch (error.code) {
+		case 'EACCES':
+			console.error(bind + ' requires elevated privileges');
+			process.exit(1);
+		case 'EADDRINUSE':
+			console.error(bind + ' is already in use');
+			process.exit(1);
+		default:
+			throw error;
+	}
+};
 
 module.exports = errorHandler;
