+++ api/helpers/normalizePort.js
@@ -1,16 +1,15 @@
 const normalizePort = (val) => {
-    const port = parseInt(val, 10);
-  
-    if (isNaN(port)) {
-      return val;
-    }
-  
-    if (port >= 0) {
-      return port;
-    }
-  
-    return false;
-}
+	const port = parseInt(val, 10);
+
+	if (isNaN(port)) {
+		return val;
+	}
+
+	if (port >= 0) {
+		return port;
+	}
+
+	return false;
+};
 
 module.exports = normalizePort;
-
