--- api/helpers/onError.js
@@ -1,24 +1,0 @@
-const onError = (error) => {
-    if (error.syscall !== 'listen') {
-      throw error;
-    }
-  
-    const bind = typeof port === 'string'
-      ? 'Pipe ' + port
-      : 'Port ' + port;
-  
-    switch (error.code) {
-      case 'EACCES':
-        console.error(bind + ' requires elevated privileges');
-        process.exit(1);
-        break;
-      case 'EADDRINUSE':
-        console.error(bind + ' is already in use');
-        process.exit(1);
-        break;
-      default:
-        throw error;
-    }
-}
-
-module.exports = onError;
